refactoring (FIN-4) move update_image_src_in_database from tasks.py to utils.py
