+++ src/media/tasks.py
@@ -3,10 +3,9 @@ from typing import AnyStr
 
 from loguru import logger
 
-from src.celery_init import celery_app, scoped_session
+from src.celery_init import celery_app
 from src.media import config as media_config
-from src.media.service import update_media_src
-from src.media.utils import process_image
+from src.media.utils import process_image, update_image_src_in_database
 from src.s3.service import send_file_to_storage
 
 
@@ -35,14 +34,3 @@ def load_image(image_id: int, image_data: AnyStr) -> None:
     asyncio.get_event_loop().run_until_complete(
         update_image_src_in_database(image_id=image_id, image_src=image_src)
     )
-
-
-async def update_image_src_in_database(image_id: int, image_src: str) -> None:
-    """
-    async update image src in the database.
-    :param image_id: the image record id.
-    :param image_src: new image src in storage
-    :return: None
-    """
-    async with scoped_session() as session:
-        await update_media_src(media_id=image_id, media_src=image_src, session=session)
